Added support for multiple layers of clouds to planets. Implemented for Cytherean (Venus-like) worlds, which have thick cloud cover, giving them two different layers of slightly different colours, which requires coloured transparency maps.
Significant updates to how worlds are displayed as 3D globes. Their rotation can
now be controlled with the mouse, and moving the planet around temporarily halts
the auto rotation, and makes the clouds transparent. When you stop moving the
mouse, the clouds fade back in and then the rotation begins again. This allows
you to see the surface of worlds covered in thick clouds.

We may not want this effect on Jovian worlds, so need to consider that later.

+++ src/main/java/uk/org/glendale/worldgen/astro/planets/generators/terrestrial/EoGaian.java
@@ -24,7 +24,7 @@ import static uk.org.glendale.worldgen.astro.commodities.CommodityName.*;
 public class EoGaian extends Terrestrial {
     private static final Logger logger = LoggerFactory.getLogger(EoGaian.class);
 
-    public EoGaian(WorldGen worldgen, StarSystem system, Star star, Planet previous, int distance) {
+    public EoGaian(WorldGen worldgen, StarSystem system, Star star, Planet previous, long distance) {
         super(worldgen, system, star, previous, distance);
     }
